rating routes and global adjustments

+++ content/content_routes.py
@@ -4,15 +4,15 @@ from db.mongodb_helpers import *
 import logging
 from content.content_controller import ContentController
 from models.content import *
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
-content_routes = Blueprint("contents_bp", __name__)
+content_routes = Blueprint("content_bp", __name__)
 content_controller = ContentController()
 content_logger = logging.getLogger(__name__)
 content_logger.setLevel(logging.DEBUG)
 
+
 @content_routes.route("/content/<string:content_id>", methods=["GET"])
-@jwt_required()
 def get_content(content_id: str) -> ContentModel:
     try:
         content_logger.info(f"get_content called with url {request.url}")
@@ -32,7 +32,6 @@ def get_content(content_id: str) -> ContentModel:
 
 
 @content_routes.route("/content", methods=["GET", "POST"])
-@jwt_required()
 def get_contents() -> ListSuccessResponse:
     try:
         content_logger.info(f"get_content called with url {request.url}")
@@ -59,15 +58,18 @@ def get_contents() -> ListSuccessResponse:
         content_logger.error(f"Exception: {str(e)}")
         return jsonify({"error": str(e)}), 500
 
+
 @content_routes.route("/content/add", methods=["POST"])
 @jwt_required()
 def add_content() -> ContentModel:
     try:
         content_logger.info(f"add_content called with url {request.url}")
+        current_user = get_jwt_identity()
+        request.json["created_by"] = current_user
+        content_logger.debug(request.json)
         content_input = ContentInputModel(**request.json)
         content = content_controller.add_content(content_input)
         return content.to_json(), 201
-    
     except ValueError as e:
         content_logger.error(f"Value Error: {str(e)}")
         return jsonify({"error": str(e)}), 400
@@ -77,23 +79,37 @@ def add_content() -> ContentModel:
     except Exception as e:
         content_logger.error(f"Exception: {str(e)}")
         return jsonify({"error": str(e)}), 500
-    
+
+
 @content_routes.route("/content/<string:content_id>", methods=["PUT"])
 @jwt_required()
 def update_content(content_id: str) -> Dict[str, Any]:
     try:
         content_logger.info(f"content called with url {request.url}")
         existing_content = content_controller.get_content(content_id)
+        current_user = get_jwt_identity()
+
+        if existing_content.created_by != current_user:
+            return (
+                jsonify(
+                    {
+                        "error": "You are not allowed to edit content that you did not create"
+                    }
+                ),
+                403,
+            )
 
         if request.json == {}:
             content_logger.info(f"Empty request body")
             return jsonify({"error": "Empty request body"}), 400
         update_data = PartialContentModel(**request.json)
-        
+
         if existing_content is None:
             content_logger.info(f"Content not found")
             return jsonify({"error": "Content not found"}), 404
-        updated_content = content_controller.update_content(update_data.dict(exclude_unset=True), content_id)
+        updated_content = content_controller.update_content(
+            update_data.dict(exclude_unset=True), content_id
+        )
 
         return updated_content.to_json(), 200
 
@@ -104,16 +120,28 @@ def update_content(content_id: str) -> Dict[str, Any]:
         content_logger.error(f"Exception: {str(e)}")
         return jsonify({"error": str(e)}), 500
 
+
 @content_routes.route("/content/<string:content_id>", methods=["DELETE"])
+@jwt_required()
 def delete_content(content_id: str):
     try:
         content_logger.info(f"delete_content called with url {request.url}")
         existing_content = content_controller.get_content(content_id)
-        
+        current_user = get_jwt_identity()
+        if existing_content.created_by != current_user:
+            return (
+                jsonify(
+                    {
+                        "error": "You are not allowed to delete content that you did not create"
+                    }
+                ),
+                403,
+            )
+
         if existing_content is None:
             content_logger.info(f"Content not found")
             return jsonify({"error": "Content not found"}), 404
-        
+
         content_controller.delete_content(content_id)
 
         return jsonify({"message": "Content deleted successfully"}), 200
